Perform when basket is over capacity
In some cases the basket might end up with more entries than the `queue_class.basket_options_hash[:size]` which prevents the basket from ever performing/or being cleared.

These scenarios could happen when a certain record is added to the queue_collection but the `on_success` or `perform` raises an exception prior to the basket being cleared — and a subsequent `Basket.add` pushes the basket capacity passed the defined size.

In this scenario ^, we likely want to ensure the basket can be cleared/performed when the underlying issue that raised an exception is addressed.